backend/myaudio/requirement.txt
Django==4.0.1
pillow
django-cors-headers==3.11.0
djangorestframework>=3.13,<4
graphene>=2.1,<3
graphene-django>=2.1,<3
graphql-core>=2.1,<3
graphql-relay==2.0.1